facerecog_tensorflow_gcp/gcf_save_image/requirements.txt
oauth2client
google-api-python-client
opencv-python
pillow
numpy
google-cloud-storage

